requirements.txt
python-telegram-bot==22.1
aiosqlite==0.19.0
